Add button to create CatFact (not implemented)
At the moment the button creates an alert when clicked. I started to
implement k8s resource creation but realized that k8sCreateResource is
not available in any of the packages in package.json. Turns out the type
definitions for k8sCreateResource exist but not the implementation.

Further, I found that the dynamic plugin SDK that is in package.json
isn't the current version. They are now released under @openshift
instead of @openshift-console.